get rid of 64bit warnings

+++ src/models/src/NITableViewSystem.h
@@ -78,7 +78,7 @@
 - (BOOL)reloadCellForTableItem:(id)object withRowAnimation:(UITableViewRowAnimation)rowAnimation;
 - (BOOL)reloadCellsForTableItems:(NSArray *)objects withRowAnimation:(UITableViewRowAnimation)rowAnimation;
 
-- (void)insertTableItem:(id)object atIndex:(int)index withRowAnimation:(UITableViewRowAnimation)animation;
+- (void)insertTableItem:(id)object atIndex:(NSUInteger)index withRowAnimation:(UITableViewRowAnimation)animation;
 - (void)insertTableItem:(id)object atIndexPath:(NSIndexPath *)indexPath withRowAnimation:(UITableViewRowAnimation)animation;
 
 /**
